Misc bugfixes around some not so often used features (#65)
* Fix bullets not traveling over lava

* Add bullet element tests

* Fix OOP tracing messages when labels are executed

* Add a test for :SHOT label

* Fix messages not clearing when changing boards on Super

* Make the fade effect work on Super

* Refactor fade matrix to its own unit

* Optimizations
- add HWord
- add functions that replace LINQ queries against lists
- allow fixed lists to start from an index other than 0
- remove unused code

* Small optimization in element - no need to convert the whole "code edit text" content when the engine only needs to know if there's anything there at all

* MusicEncoder now accepts char span

* Move `Context` out to root namespace

* Namespace simplification

* Fix sound, add hud error messages for wrong version loaded instead of outright crashing

* Fix bugs:
- loading a world didn't properly change the player to a monitor on the title screen
- webs are sticky now
- adjacent elements during drawing now count the board edge

* Fix player movement vector

* Fix an issue where the monitor element was causing the game loop to break in odd ways and swallow input

* Add a "no pester" flag that will disable some sounds and built-in alerts

* Fix board timer

+++ Source/Roton.Test/Roton/Integration/Oop/EndToEndTests.cs
@@ -2,7 +2,6 @@ using System.Linq;
 using AwesomeAssertions;
 using NUnit.Framework;
 using Roton.Emulation.Core.Impl;
-using Roton.Emulation.Data.Impl;
 
 namespace Roton.Test.Roton.Integration.Oop;
 
@@ -94,10 +93,10 @@ public class EndToEndTests(Context context) : OopTestFixture(context)
         Type(AnsiKey.Up);
         StepAllKeys();
 
-        actors[0].P1.Should().Be(14);
-        actors[1].P1.Should().Be(14);
-        actors[2].P1.Should().Be(14);
-        actors[3].P1.Should().Be(13);
-        actors[4].P1.Should().Be(14);
+        ((int)actors[0].P1).Should().Be(14);
+        ((int)actors[1].P1).Should().Be(14);
+        ((int)actors[2].P1).Should().Be(14);
+        ((int)actors[3].P1).Should().Be(13);
+        ((int)actors[4].P1).Should().Be(14);
     }
 }
